Consolidated class-specific functions for TrajectoryData and ShotProps. Moved ShotProps to the conditions module.
