+++ py_ballisticcalc/trajectory_data.py
@@ -5,7 +5,6 @@ Core Components:
     - TrajectoryData: Detailed ballistic state at a single trajectory point
     - BaseTrajData: Minimal trajectory data required for a single point
     - HitResult: Complete trajectory results with metadata
-    - ShotProps: All metadata used to compute the trajectory
     - DangerSpace: Target engagement zone analysis
 
 Key Features:
@@ -52,17 +51,12 @@ See Also:
 from __future__ import annotations
 import math
 import typing
-import warnings
 from dataclasses import dataclass, field
 from deprecated import deprecated
-from typing_extensions import Final, List, Literal, NamedTuple, Optional, Sequence, Tuple, Union
+from typing_extensions import Final, Literal, NamedTuple, Optional, Tuple, Union
 
-from py_ballisticcalc.conditions import Shot, Wind
-from py_ballisticcalc.drag_model import DragDataPoint
 from py_ballisticcalc.exceptions import RangeError
-from py_ballisticcalc.unit import (Angular, Distance, Energy, Pressure, Temperature, Velocity, Weight,
-                                   GenericDimension, Unit, PreferredUnits
-)
+from py_ballisticcalc.unit import Angular, Distance, Energy, Velocity, Weight, GenericDimension, Unit, PreferredUnits
 from py_ballisticcalc.vector import Vector
 from py_ballisticcalc.interpolation import (
     InterpolationMethod,
@@ -73,20 +67,14 @@ from py_ballisticcalc.interpolation import (
 if typing.TYPE_CHECKING:
     from pandas import DataFrame
     from matplotlib.axes import Axes
+    from py_ballisticcalc.conditions import ShotProps
 
 __all__ = (
     'TrajFlag',
     'BaseTrajData',
     'TrajectoryData',
-    'ShotProps',
     'HitResult',
     'DangerSpace',
-    'CurvePoint',
-    'create_trajectory_row',
-    'calculate_energy',
-    'calculate_ogw',
-    'get_correction',
-    'calculate_curve',
 )
 
 
@@ -197,22 +185,6 @@ class TrajFlag(int):
             parts.remove("ZERO_UP")
             parts.remove("ZERO_DOWN")
         return "|".join(parts) if parts else "UNKNOWN"
-
-
-# Interpolation helpers moved to py_ballisticcalc.interpolation
-
-
-class CurvePoint(NamedTuple):
-    """Coefficients for quadratic curve fitting.
-    
-    Attributes:
-        a: Quadratic coefficient (x² term) in the equation y = ax² + bx + c.
-        b: Linear coefficient (x term) in the equation y = ax² + bx + c.
-        c: Constant coefficient (constant term) in the equation y = ax² + bx + c.
-    """
-    a: float
-    b: float
-    c: float
 
 
 class BaseTrajData(NamedTuple):
@@ -431,6 +403,82 @@ class TrajectoryData(NamedTuple):
         return TrajectoryData.from_props(props, data.time, data.position, data.velocity, data.mach, flag)
 
     @staticmethod
+    def get_correction(distance: float, offset: float) -> float:
+        """Calculates the sight adjustment in radians.
+
+        Args:
+            distance: The distance to the target in feet.
+            offset: The offset from the target in feet.
+
+        Returns:
+            The sight adjustment in radians.
+        """
+        if distance != 0:
+            return math.atan(offset / distance)
+        return 0  # None
+
+    @staticmethod
+    def calculate_energy(bullet_weight: float, velocity: float) -> float:
+        """Calculates the kinetic energy of a projectile.
+
+        Args:
+            bullet_weight: The weight of the bullet in pounds.
+            velocity: The velocity of the bullet in feet per second.
+
+        Returns:
+            The kinetic energy of the projectile in foot-pounds.
+        """
+        return bullet_weight * math.pow(velocity, 2) / 450400
+
+    @staticmethod
+    def calculate_ogw(bullet_weight: float, velocity: float) -> float:
+        """Calculates the optimal game weight for a projectile.
+
+        Args:
+            bullet_weight: The weight of the bullet in pounds.
+            velocity: The velocity of the bullet in feet per second.
+
+        Returns:
+            The optimal game weight in pounds.
+        """
+        return math.pow(bullet_weight, 2) * math.pow(velocity, 3) * 1.5e-12
+
+    @staticmethod
+    def _new_feet(v: float):
+        d = object.__new__(Distance)
+        d._value = v * 12
+        d._defined_units = Unit.Foot
+        return d
+
+    @staticmethod
+    def _new_fps(v: float):
+        d = object.__new__(Velocity)
+        d._value = v / 3.2808399
+        d._defined_units = Unit.FPS
+        return d
+
+    @staticmethod
+    def _new_rad(v: float):
+        d = object.__new__(Angular)
+        d._value = v
+        d._defined_units = Unit.Radian
+        return d
+
+    @staticmethod
+    def _new_ft_lb(v: float):
+        d = object.__new__(Energy)
+        d._value = v
+        d._defined_units = Unit.FootPound
+        return d
+
+    @staticmethod
+    def _new_lb(v: float):
+        d = object.__new__(Weight)
+        d._value = v / 0.000142857143
+        d._defined_units = Unit.Pound
+        return d
+
+    @staticmethod
     def from_props(props: ShotProps,
                     time: float,
                     range_vector: Vector,
@@ -441,8 +489,8 @@ class TrajectoryData(NamedTuple):
         spin_drift = props.spin_drift(time)
         velocity = velocity_vector.magnitude()
         windage = range_vector.z + spin_drift
-        drop_adjustment = get_correction(range_vector.x, range_vector.y)
-        windage_adjustment = get_correction(range_vector.x, windage)
+        drop_adjustment = TrajectoryData.get_correction(range_vector.x, range_vector.y)
+        windage_adjustment = TrajectoryData.get_correction(range_vector.x, windage)
         trajectory_angle = math.atan2(velocity_vector.y, velocity_vector.x)
         look_angle_cos = math.cos(props.look_angle_rad)
         look_angle_sin = math.sin(props.look_angle_rad)
@@ -450,20 +498,20 @@ class TrajectoryData(NamedTuple):
         drag = props.drag_by_mach(velocity / mach)
         return TrajectoryData(
             time=time,
-            distance=_new_feet(range_vector.x),
-            velocity=_new_fps(velocity),
+            distance=TrajectoryData._new_feet(range_vector.x),
+            velocity=TrajectoryData._new_fps(velocity),
             mach=velocity / mach,
-            height=_new_feet(range_vector.y),
-            slant_height=_new_feet(range_vector.y * look_angle_cos - range_vector.x * look_angle_sin),
-            drop_adj=_new_rad(drop_adjustment - (props.look_angle_rad if range_vector.x else 0)),
-            windage=_new_feet(windage),
-            windage_adj=_new_rad(windage_adjustment),
-            slant_distance=_new_feet(range_vector.x * look_angle_cos + range_vector.y * look_angle_sin),
-            angle=_new_rad(trajectory_angle),
+            height=TrajectoryData._new_feet(range_vector.y),
+            slant_height=TrajectoryData._new_feet(range_vector.y * look_angle_cos - range_vector.x * look_angle_sin),
+            drop_adj=TrajectoryData._new_rad(drop_adjustment - (props.look_angle_rad if range_vector.x else 0)),
+            windage=TrajectoryData._new_feet(windage),
+            windage_adj=TrajectoryData._new_rad(windage_adjustment),
+            slant_distance=TrajectoryData._new_feet(range_vector.x * look_angle_cos + range_vector.y * look_angle_sin),
+            angle=TrajectoryData._new_rad(trajectory_angle),
             density_ratio=density_ratio,
             drag=drag,
-            energy=_new_ft_lb(calculate_energy(props.weight_grains, velocity)),
-            ogw=_new_lb(calculate_ogw(props.weight_grains, velocity)),
+            energy=TrajectoryData._new_ft_lb(TrajectoryData.calculate_energy(props.weight_grains, velocity)),
+            ogw=TrajectoryData._new_lb(TrajectoryData.calculate_ogw(props.weight_grains, velocity)),
             flag=flag
         )
 
@@ -553,430 +601,6 @@ class TrajectoryData(NamedTuple):
 
         interpolated_fields['flag'] = flag
         return TrajectoryData(**interpolated_fields)
-
-
-@dataclass
-class ShotProps:
-    """Shot configuration and parameters for ballistic trajectory calculations.
-    
-    Contains all shot-specific data converted to internal units for high-performance
-    ballistic calculations. This class serves as the computational interface between
-    user-friendly Shot objects and the numerical integration engines.
-    
-    The class pre-computes expensive calculations (ballistic coefficient curves,
-    atmospheric data, projectile properties) and stores them in optimized formats
-    for repeated use during trajectory integration. All values are converted to
-    internal units (feet, seconds, grains) for computational efficiency.
-        
-    Examples:
-        Create ShotProps from Shot object:
-        
-        ```python
-        from py_ballisticcalc import Shot, ShotProps
-        
-        # Create shot configuration
-        shot = Shot(weapon=weapon, ammo=ammo, atmo=atmo)
-
-        # Convert to ShotProps
-        shot_props = ShotProps.from_shot(shot)
-        
-        # Access pre-computed values
-        print(f"Stability coefficient: {shot_props.stability_coefficient}")
-
-        # Get drag coefficient at specific Mach number
-        drag = shot_props.drag_by_mach(1.5)
-        
-        # Calculate spin drift at flight time
-        time = 1.2  # seconds
-        drift = shot_props.spin_drift(time)  # inches
-        
-        # Get atmospheric conditions at altitude
-        drop = 100  # feet above initial altitude
-        density_ratio, mach_fps = shot_props.get_density_and_mach_for_altitude(drop)
-        ```
-        
-    Computational Optimizations:
-        - Drag coefficient curves pre-computed for fast interpolation
-        - Trigonometric values (cant_cosine, cant_sine) pre-calculated
-        - Atmospheric parameters cached for repeated altitude lookups
-        - Miller stability coefficient computed once during initialization
-        
-    Note:
-        This class is designed for internal use by ballistic calculation engines.
-        User code should typically work with Shot objects and let the Calculator
-        handle the conversion to ShotProps automatically.
-        
-        The original Shot object is retained for reference, but modifications
-        to it after ShotProps creation will not affect the stored calculations.
-        Create a new ShotProps instance if Shot parameters change.
-
-    TODO: The Shot member object should either be a copy or immutable so that subsequent changes to its
-          properties do not invalidate the calculations and data associated with this ShotProps instance.
-    """
-    shot: Shot  # Reference to the original Shot object
-    bc: float  # Ballistic coefficient
-    curve: List[CurvePoint]  # Pre-computed drag curve points
-    mach_list: List[float]  # List of Mach numbers for interpolation
-
-    look_angle_rad: float  # Slant angle in radians
-    twist_inch: float  # Twist rate of barrel rifling, in inches of length to make one full rotation
-    length_inch: float  # Length of the bullet in inches
-    diameter_inch: float  # Diameter of the bullet in inches
-    weight_grains: float  # Weight of the bullet in grains
-    barrel_elevation_rad: float  # Barrel elevation angle in radians
-    barrel_azimuth_rad: float  # Barrel azimuth angle in radians
-    sight_height_ft: float  # Height of the sight above the bore in feet
-    cant_cosine: float  # Cosine of the cant angle
-    cant_sine: float  # Sine of the cant angle
-    alt0_ft: float  # Initial altitude in feet
-    muzzle_velocity_fps: float  # Muzzle velocity in feet per second
-    stability_coefficient: float = field(init=False)  # Miller stability coefficient
-    calc_step: float = field(init=False)  # Calculation step size
-    filter_flags: Union[TrajFlag, int] = field(init=False)  # Flags for special ballistic trajectory points
-
-    def __post_init__(self):
-        self.stability_coefficient = self._calc_stability_coefficient()
-
-    @property
-    def winds(self) -> Sequence[Wind]:
-        return self.shot.winds
-
-    @property
-    def look_angle(self) -> Angular:
-        return Angular.Radian(self.look_angle_rad)
-
-    @classmethod
-    def from_shot(cls, shot: Shot) -> ShotProps:
-        """Initializes a ShotProps instance from a Shot instance."""
-        return cls(
-            shot=shot,
-            bc=shot.ammo.dm.BC,
-            curve=calculate_curve(shot.ammo.dm.drag_table),
-            mach_list=_get_only_mach_data(shot.ammo.dm.drag_table),
-            look_angle_rad=shot.look_angle >> Angular.Radian,
-            twist_inch=shot.weapon.twist >> Distance.Inch,
-            length_inch=shot.ammo.dm.length >> Distance.Inch,
-            diameter_inch=shot.ammo.dm.diameter >> Distance.Inch,
-            weight_grains=shot.ammo.dm.weight >> Weight.Grain,
-            barrel_elevation_rad=shot.barrel_elevation >> Angular.Radian,
-            barrel_azimuth_rad=shot.barrel_azimuth >> Angular.Radian,
-            sight_height_ft=shot.weapon.sight_height >> Distance.Foot,
-            cant_cosine=math.cos(shot.cant_angle >> Angular.Radian),
-            cant_sine=math.sin(shot.cant_angle >> Angular.Radian),
-            alt0_ft=shot.atmo.altitude >> Distance.Foot,
-            muzzle_velocity_fps=shot.ammo.get_velocity_for_temp(shot.atmo.powder_temp) >> Velocity.FPS,
-        )
-
-    def drag_by_mach(self, mach: float) -> float:
-        """Calculates a standard drag factor (SDF) for the given Mach number:
-            Drag force = V^2 * AirDensity * C_d * S / 2m
-                       = V^2 * density_ratio * SDF
-        Where:
-            - density_ratio = LocalAirDensity / StandardDensity = rho / rho_0
-            - StandardDensity of Air = rho_0 = 0.076474 lb/ft^3
-            - S is cross-section = d^2 pi/4, where d is bullet diameter in inches
-            - m is bullet mass in pounds
-            - bc contains m/d^2 in units lb/in^2, which is multiplied by 144 to convert to lb/ft^2
-        Thus:
-            - The magic constant found here = StandardDensity * pi / (4 * 2 * 144)
-
-        Args:
-            mach: The Mach number.
-
-        Returns:
-            The standard drag factor at the given Mach number.
-        """
-        # cd = calculate_by_curve(self._table_data, self._curve, mach)
-        # use calculation over list[double] instead of list[DragDataPoint]
-        cd = _calculate_by_curve_and_mach_list(self.mach_list, self.curve, mach)
-        return cd * 2.08551e-04 / self.bc
-
-    def spin_drift(self, time) -> float:
-        """Litz spin-drift approximation
-
-        Args:
-            time: Time of flight
-
-        Returns:
-            float: Windage due to spin drift, in inches
-        """
-        if (self.stability_coefficient != 0) and (self.twist_inch != 0):
-            sign = 1 if self.twist_inch > 0 else -1
-            return sign * (1.25 * (self.stability_coefficient + 1.2)
-                           * math.pow(time, 1.83)) / 12
-        return 0
-
-    def _calc_stability_coefficient(self) -> float:
-        """Calculates the Miller stability coefficient.
-
-        Returns:
-            float: The Miller stability coefficient.
-        """
-        if self.twist_inch and self.length_inch and self.diameter_inch and self.shot.atmo.pressure.raw_value:
-            twist_rate = math.fabs(self.twist_inch) / self.diameter_inch
-            length = self.length_inch / self.diameter_inch
-            # Miller stability formula
-            sd = 30 * self.weight_grains / (
-                    math.pow(twist_rate, 2) * math.pow(self.diameter_inch, 3) * length * (1 + math.pow(length, 2))
-            )
-            # Velocity correction factor
-            fv = math.pow(self.muzzle_velocity_fps / 2800, 1.0 / 3.0)
-            # Atmospheric correction
-            ft = self.shot.atmo.temperature >> Temperature.Fahrenheit
-            pt = self.shot.atmo.pressure >> Pressure.InHg
-            ftp = ((ft + 460) / (59 + 460)) * (29.92 / pt)
-            return sd * fv * ftp
-        return 0
-
-    def get_density_and_mach_for_altitude(self, drop: float) -> Tuple[float, float]:
-        """Gets the air density and Mach number for a given altitude.
-
-        Args:
-            drop: The change in feet from the initial altitude.
-
-        Returns:
-            A tuple containing the air density (in lb/ft³) and Mach number at the specified altitude.
-        """
-        return self.shot.atmo.get_density_and_mach_for_altitude(self.alt0_ft + drop)
-
-
-# pylint: disable=too-many-positional-arguments
-@deprecated(reason="Use TrajectoryData.from_props instead.")
-def create_trajectory_row(time: float, range_vector: Vector, velocity_vector: Vector,
-                          velocity: float, mach: float, spin_drift: float, look_angle: float,
-                          density_ratio: float, drag: float, weight: float,
-                          flag: Union[TrajFlag, int]) -> TrajectoryData:
-    """Creates a TrajectoryData object representing a single row of trajectory data.
-
-    Args:
-        time: Time of flight in seconds.
-        range_vector: Position vector in feet.
-        velocity_vector: Velocity vector in fps.
-        velocity: Velocity magnitude in fps.
-        mach: Mach number.
-        spin_drift: Spin drift in feet.
-        look_angle: Slant angle in radians.
-        density_ratio: Density ratio (rho / rho_0).
-        drag: Drag value.
-        weight: Weight value.
-        flag: Flag value.
-
-    Returns:
-        A TrajectoryData object representing the trajectory data.
-    """
-    warnings.warn("This method is deprecated", DeprecationWarning)
-
-    windage = range_vector.z + spin_drift
-    drop_adjustment = get_correction(range_vector.x, range_vector.y)
-    windage_adjustment = get_correction(range_vector.x, windage)
-    trajectory_angle = math.atan2(velocity_vector.y, velocity_vector.x)
-
-    return TrajectoryData(
-        time=time,
-        distance=_new_feet(range_vector.x),
-        velocity=_new_fps(velocity),
-        mach=velocity / mach,
-        height=_new_feet(range_vector.y),
-        slant_height=_new_feet(range_vector.y * math.cos(look_angle) - range_vector.x * math.sin(look_angle)),
-        drop_adj=_new_rad(drop_adjustment - (look_angle if range_vector.x else 0)),
-        windage=_new_feet(windage),
-        windage_adj=_new_rad(windage_adjustment),
-        slant_distance=_new_feet(range_vector.x * math.cos(look_angle) + range_vector.y * math.sin(look_angle)),
-        angle=_new_rad(trajectory_angle),
-        density_ratio=density_ratio,
-        drag=drag,
-        energy=_new_ft_lb(calculate_energy(weight, velocity)),
-        ogw=_new_lb(calculate_ogw(weight, velocity)),
-        flag=flag
-    )
-
-
-def _new_feet(v: float):
-    d = object.__new__(Distance)
-    d._value = v * 12
-    d._defined_units = Unit.Foot
-    return d
-
-
-def _new_fps(v: float):
-    d = object.__new__(Velocity)
-    d._value = v / 3.2808399
-    d._defined_units = Unit.FPS
-    return d
-
-
-def _new_rad(v: float):
-    d = object.__new__(Angular)
-    d._value = v
-    d._defined_units = Unit.Radian
-    return d
-
-
-def _new_ft_lb(v: float):
-    d = object.__new__(Energy)
-    d._value = v
-    d._defined_units = Unit.FootPound
-    return d
-
-
-def _new_lb(v: float):
-    d = object.__new__(Weight)
-    d._value = v / 0.000142857143
-    d._defined_units = Unit.Pound
-    return d
-
-
-def get_correction(distance: float, offset: float) -> float:
-    """Calculates the sight adjustment in radians.
-
-    Args:
-        distance: The distance to the target in feet.
-        offset: The offset from the target in feet.
-
-    Returns:
-        The sight adjustment in radians.
-    """
-    if distance != 0:
-        return math.atan(offset / distance)
-    return 0  # None
-
-
-def calculate_energy(bullet_weight: float, velocity: float) -> float:
-    """Calculates the kinetic energy of a projectile.
-
-    Args:
-        bullet_weight: The weight of the bullet in pounds.
-        velocity: The velocity of the bullet in feet per second.
-
-    Returns:
-        The kinetic energy of the projectile in foot-pounds.
-    """
-    return bullet_weight * math.pow(velocity, 2) / 450400
-
-
-def calculate_ogw(bullet_weight: float, velocity: float) -> float:
-    """Calculates the optimal game weight for a projectile.
-
-    Args:
-        bullet_weight: The weight of the bullet in pounds.
-        velocity: The velocity of the bullet in feet per second.
-
-    Returns:
-        The optimal game weight in pounds.
-    """
-    return math.pow(bullet_weight, 2) * math.pow(velocity, 3) * 1.5e-12
-
-
-def calculate_curve(data_points: List[DragDataPoint]) -> List[CurvePoint]:
-    """Piecewise quadratic interpolation of drag curve
-
-    Args:
-        data_points: List[{Mach, CD}] data_points in ascending Mach order
-
-    Returns:
-        List[CurvePoints] to interpolate drag coefficient
-    """
-    rate = (data_points[1].CD - data_points[0].CD) / (data_points[1].Mach - data_points[0].Mach)
-    curve = [CurvePoint(0, rate, data_points[0].CD - data_points[0].Mach * rate)]
-    len_data_points = int(len(data_points))
-    len_data_range = len_data_points - 1
-
-    for i in range(1, len_data_range):
-        x1 = data_points[i - 1].Mach
-        x2 = data_points[i].Mach
-        x3 = data_points[i + 1].Mach
-        y1 = data_points[i - 1].CD
-        y2 = data_points[i].CD
-        y3 = data_points[i + 1].CD
-        a = ((y3 - y1) * (x2 - x1) - (y2 - y1) * (x3 - x1)) / (
-            (x3 * x3 - x1 * x1) * (x2 - x1) - (x2 * x2 - x1 * x1) * (x3 - x1))
-        b = (y2 - y1 - a * (x2 * x2 - x1 * x1)) / (x2 - x1)
-        c = y1 - (a * x1 * x1 + b * x1)
-        curve_point = CurvePoint(a, b, c)
-        curve.append(curve_point)
-
-    num_points = len_data_points
-    rate = (data_points[num_points - 1].CD - data_points[num_points - 2].CD) / \
-           (data_points[num_points - 1].Mach - data_points[num_points - 2].Mach)
-    curve_point = CurvePoint(
-        0, rate, data_points[num_points - 1].CD - data_points[num_points - 2].Mach * rate
-    )
-    curve.append(curve_point)
-    return curve
-
-
-def _get_only_mach_data(data: List[DragDataPoint]) -> List[float]:
-    """Extracts Mach values from a list of DragDataPoint objects.
-
-    Args:
-        data: A list of DragDataPoint objects.
-
-    Returns:
-        A list containing only the Mach values from the input data.
-    """
-    return [dp.Mach for dp in data]
-
-# # use ._get_only_mach_data with ._calculate_by_curve_and_mach_list because it's faster
-# def calculate_by_curve(data: List[DragDataPoint], curve: List[CurvePoint], mach: float) -> float:
-#     """
-#     Binary search for drag coefficient based on Mach number
-#     :param data: data
-#     :param curve: Output of calculate_curve(data)
-#     :param mach: Mach value for which we're searching for CD
-#     :return float: drag coefficient
-#     """
-#     num_points = int(len(curve))
-#     mlo = 0
-#     mhi = num_points - 2
-#
-#     while mhi - mlo > 1:
-#         mid = int(math.floor(mhi + mlo) / 2.0)
-#         if data[mid].Mach < mach:
-#             mlo = mid
-#         else:
-#             mhi = mid
-#
-#     if data[mhi].Mach - mach > mach - data[mlo].Mach:
-#         m = mlo
-#     else:
-#         m = mhi
-#     curve_m = curve[m]
-#     return curve_m.c + mach * (curve_m.b + curve_m.a * mach)
-
-
-def _calculate_by_curve_and_mach_list(mach_list: List[float], curve: List[CurvePoint], mach: float) -> float:
-    """Calculates a value based on a piecewise quadratic curve and a list of Mach values.
-
-    This function performs a binary search on the `mach_list` to find the segment
-    of the `curve` relevant to the input `mach` number and then interpolates
-    the value using the quadratic coefficients of that curve segment.
-
-    Args:
-        mach_list: A sorted list of Mach values corresponding to the `curve` points.
-        curve: A list of CurvePoint objects, where each object
-            contains quadratic coefficients (a, b, c) for a Mach number segment.
-        mach: The Mach number at which to calculate the value.
-
-    Returns:
-        The calculated value based on the interpolated curve at the given Mach number.
-    """
-    num_points = len(curve)
-    mlo = 0
-    mhi = num_points - 2
-
-    while mhi - mlo > 1:
-        mid = (mhi + mlo) // 2
-        if mach_list[mid] < mach:
-            mlo = mid
-        else:
-            mhi = mid
-
-    if mach_list[mhi] - mach > mach - mach_list[mlo]:
-        m = mlo
-    else:
-        m = mhi
-    curve_m = curve[m]
-    return curve_m.c + mach * (curve_m.b + curve_m.a * mach)
 
 
 class DangerSpace(NamedTuple):
